requirements.txt
pyTelegramBotAPI==4.14.0
websocket-client==1.7.0
requests==2.31.0
pydub==0.25.1
python-dotenv==1.0.0
